requirements.txt
requests
python-dotenv
langchain
langchain_openai
openai
matplotlib
networkx
docx
docling
kor
tqdm
aiofiles
